fix: 🐛 verify
removing file lookup since files may not exist until after exec
